+++ proto/client-api/CHANGELOG.md
@@ -6,15 +6,24 @@ All user visible changes to this project will be documented in this file. This p
 
 
 
-## TBD [0.1.1] · 2019-??-??
-[0.1.1]: /../../tree/medea-client-api-proto-0.1.1/proto/client-api
+## TBD [0.2.0] · 2019-??-??
+[0.2.0]: /../../tree/medea-client-api-proto-0.2.0/proto/client-api
+
+### BC Breaks
+
+- RPC messages ([#75](/../../pull/75)):
+    - Server messages:
+        - `Pong` is now `Ping`.
+    - Client messages:
+        - `Ping` is now `Pong`.
 
 ### Added
 
 - `TrackId` and `PeerId` types ([#28]);
 - `Incrementable` trait ([#28]);
 - `CloseReason` and `CloseDescription` types ([#58](/../../pull/58));
-- `AddPeerConnectionMetrics` client command with `IceConnectionState` metric ([#71](/../../pull/71)).
+- `AddPeerConnectionMetrics` client command with `IceConnectionState` metric ([#71](/../../pull/71));
+- `RpcSettings` server message ([#75](/../../pull/75)).
 
 [#28]: /../../pull/28